wrapped the bot in a docker container
