+++ bot/README.md
@@ -1,9 +1,10 @@
-Данный бот имеент 3 обработчика сообщения.
-/start - отправляет пользователю приветсвенное сообщение.
-/history - отправляет пользователю его последние 5 сообщений с указанием даты отправки.
-Каждое другое сообщение записывается в базу данных.
-Для запуска бота необходимо установить все необходимые библиотеки (pip install -r requirements.txt).
-Создать файл .env в котором указать два параметра:
+Данный бот имеент 3 обработчика сообщения.  
+/start - отправляет пользователю приветсвенное сообщение.  
+/history - отправляет пользователю его последние 5 сообщений с указанием даты отправки.  
+Каждое другое сообщение записывается в базу данных.  
+Для запуска бота необходимо создать файл .env в директории bot в котором указать следующие параметры:
 TOKEN= - токен запускаемого бота.
-PG_LINK= - ссылка на бд в формате postgresql://login:password@host:password/db.
-После чего запустить файл aiogram_run.py
+POSTGRES_DB= - имя базы данных
+POSTGRES_USER= - имя пользователя
+POSTGRES_PASSWORD= - пароль базы данных
+После чего запустить выполнить команду sudo docker compose up
