+++ bot/requirements.txt
@@ -1,3 +1,3 @@
-asyncpg
-aiogram
-python-decouple
+asyncpg==0.29.0
+aiogram==3.5.0
+python-decouple==3.8
